+++ lib/src/modular_base.dart
@@ -266,7 +266,7 @@ class Modular {
 
     realGuards
         .expand((c) => c.executors)
-        .forEach((c) => c.onGuarded(path, guard == null));
+        .forEach((c) => c.onGuarded(path, isActive: guard == null));
 
     if (guard != null) {
       throw ModularError("Path guarded : $path");